feat: initialized backend
Initialized the following:

- controllers
- models
- routes

The routes for the controllers are mostly complete.

I have added "TODO" marks on sections that should be implemented later.
For the most part, you will only need to create the controllers, and the
schemas.

I have not yet initialized the authentication and athorization of the
api, just the simple routing. There is no distinction yet between a
normal user and an admin.

You can follow the Order and Transaction models for designing the
schemas.

Thank you! 🥰